python-vispy/requirements.txt
# WorldWind Python - VisPy 3D Globe Visualization
# Requirements file for pip installation

# Core visualization library
vispy>=0.14.0

# Numerical computing
numpy>=1.21.0

# Image processing (for texture loading)
Pillow>=9.0.0

# Optional: For enhanced performance
PyOpenGL>=3.1.5

# Optional: For data processing
scipy>=1.7.0

# Optional: For network requests (downloading Earth textures)
requests>=2.28.0

# Optional: For GUI framework (Qt backend for VisPy)
PyQt5>=5.15.0

# Alternative: PyQt6 (choose one)
# PyQt6>=6.0.0

# Alternative: PySide6 (choose one)
# PySide6>=6.0.0
